refactor:#15 SideSheet 컴포넌트 코드 모듈화
